BAH-4741 | Refactor. Move DataTable component to organism level
