+++ packages/bahmni-design-system/src/organisms/dataTable/index.ts
