+++ packages/bahmni-design-system/src/organisms/dataTable/types.ts
